Add editor slice for components
